Added error handling and changed the text file abit. The only bug left to handle is the blocks.list location
