+++ blocks/README.txt
@@ -1,7 +1,7 @@
 Submitted by:
-מאיה גבע(): 322854415
-ויקטוריה מוסייקו (): 322357765
-גל רובינשטיין (): 209385509
+מאיה גבע(Maya Geva): 322854415
+ויקטוריה מוסייקו (Victoria): 322357765
+גל רובינשטיין (Gal Rubinstein): 209385509
 
 INSTRUCTIONS:
 ------------
