+++ blocks/block_funcs.cpp
@@ -1,7 +1,5 @@
 #include "block_funcs.h"
-//For sleeps
-#include <thread>
-#include <chrono>
+
 #include <cstring>
 
 list<Block> readBlocksFile(const string& filename)
@@ -29,7 +27,7 @@ list<Block> readBlocksFile(const string& filename)
         else if(line.find("Time: ")==0){
             currBlock.time=line.substr(line.find(":")+2);
         }
-        else if(line.find("Relayed by: ")==0){
+        else if(line.find("Relayed By: ")==0){
             currBlock.relayed_by=line.substr(line.find(":")+2);
         }
         else if(line.find("Prev Block: ")==0){
@@ -164,107 +162,11 @@ void printValuesToCSVFile(std::string filename, std::list<Block> lst) {
     file.close();
 }
 
-void reloadDatabase(unsigned int num) {
+void reloadDatabase(unsigned int num, std::string script) {
 
-    std::string script = "../blocks.sh ";
     char tmp[5]; //Maximum amount of digits for int, especially unsigned
     sprintf(tmp, "%u", num);
     std::string numOfBlocks = std::string(tmp);
     script = script + numOfBlocks;
     system(script.c_str()); //Argument int: how many blocks will be in the list
-}
-
-void userMenu(std::list<Block>& blocks, std::string filename, std::string out) {
-
-    int choice = 0;
-
-    while (true)
-    {
-        printUserMenu();
-        std::cin >> choice;
-        performFunc(choice, blocks, filename);
-
-        if (choice == 5) //Case: refresh data
-            blocks = readBlocksFile(out);
-    }
-}
-
-void printUserMenu() {
-
-    std::cout << "Choose an option:" << std::endl;
-    std::cout << "1. Print db" << std::endl;
-    std::cout << "2. Print block by hash" << std::endl;
-    std::cout << "3. Print block by height" << std::endl;
-    std::cout << "4. Export data to csv" << std::endl;
-    std::cout << "5. Refresh data" << std::endl;
-    std::cout << "Enter your choice: ";
-}
-
-void performFunc(int choice, const std::list<Block>& blocks, std::string filename) {
-
-    switch (choice)
-    {
-    case 1:
-    {
-        printBlocks(blocks);
-        break;
-    }
-
-    case 2:
-    {
-        char tmp[70];
-        std::cout << "Enter block hash: ";
-        std::cin >> tmp;
-        std::string hash(tmp);
-        printBlockByHash(blocks, hash);
-        break;
-    }
-
-    case 3:
-    {
-        int height;
-        std::cout << "Enter block height: ";
-        std::cin >> height;
-        printBlockByHeight(blocks, height);
-        break;
-    }
-
-    case 4:
-    {
-        bool isOpen = open_new_csv_file(filename);
-        if (isOpen)
-        {
-            print_csv_menu_to_file(filename);
-            printValuesToCSVFile(filename, blocks);
-        }
-        else
-            std::cout << "Error! Could not open a new csv file!" << std::endl;
-        break;
-    }
-
-    case 5:
-    {
-        std::cout << "Please enter the number of blocks you would like to reload: ";
-        unsigned int num = 0;
-        std::cin >> num;
-        reloadDatabase(num);
-        break;
-    }
-
-    default:
-    {
-        std::cout << "Illegal choice! Printing menu again..." << std::endl;
-        break;
-    }
-
-    }
-    std::this_thread::sleep_for(std::chrono::milliseconds(500)); //Sleeps for half a second to let the user see the result before the next print
-}
-
-void loadDatabase() {
-
-    std::cout << "Please enter the number of blocks you would like to load: ";
-    unsigned int num = 0;
-    std::cin >> num;
-    reloadDatabase(num);
 }
