+++ blocks/exe1.cpp
@@ -1,10 +1,17 @@
 #include "block_funcs.h"
 
-
 int main()
 {
-    string filename = "./blocks.out";
-    list<Block> blocks = readBlocksFile(filename);
-    printBlocks(blocks);
+    try
+    {
+        string filename = "../blocks.list";
+        list<Block> blocks = readBlocksFile(filename);
+        printBlocks(blocks);
+    }
+    catch(const std::exception& e)
+    {
+        std::cout << "Could not load data: " << e.what() << std::endl;
+    }
+    
     return 0;
 }
